Move the checkbox methods to a helper
This makes the class smaller and fits in with our pattern of using
helpers for the presentation layer.
